chore(agent): 清理 Story 拆除残留死代码 + 收敛重复原语
死代码删除（全仓 grep 确认零生产引用）：
- ledger DAO 三个死读方法 listAfterSeq/countAfterSeq/findLatest（Story 拆除后无调用方，
  仅 insertMany 在用），接口 + prisma 实现一并删
- 孤儿文件 llm/http-embedding-client.ts + 其测试（零消费方，记忆系统重建时按需再建）
- context-item.utils 的 renderLlmMessagePlainText/serializeContextItem/deserializeContextItem
  及连带死掉的 renderUserMessageContent/isContextItem
- conversation-id 两个从未接线的 parse 函数、napcat 的 NapcatReceiveTextOrAtSegment/
  toStoredMessageSegments/NapcatSendGroupSegmentSchema、web-search 的 WebSearchAgentInput、
  spire 的 renderSpireUnavailable + 其 spire-unavailable.hbs 模板

陈旧注释：root-effect-interpreter 去掉已退役的 switch_state/replace_messages/focused state 字样；
kernel/contracts/llm.ts 头注释纠正为描述 LlmUsageId（原文误导为管 provider 标识）。

重复收敛：
- 抽 @kagami/shared/utils 的 BEIJING_TIME_ZONE + formatBeijingDateTime，替换 hn/ithome/
  clock/context-message-factory 四处重复的时区常量与两份逐字节相同的 formatDateTime
- 抽 apps/shared/http-retry（RETRYABLE_STATUS/computeBackoffMs/parseRetryAfter/sleep），
  替换 hn-fetch/amap-fetch 逐字节复制的退避原语

未做（PR 说明）：auth/llm-client 的 per-provider 样板判为垂直切片有意重复，wontfix；
truncate 5 处近似重复后缀/返回形状不一，收敛有改 KV 输出风险，按 issue 指引放弃。

Closes #275
