+++ apps/agent/src/agent/capabilities/ledger/infra/linear-message-ledger.dao.ts
@@ -2,11 +2,4 @@ import type { LinearMessageLedgerInsert, LinearMessageLedgerRecord } from "../do
 
 export interface LinearMessageLedgerDao {
   insertMany(entries: LinearMessageLedgerInsert[]): Promise<LinearMessageLedgerRecord[]>;
-  listAfterSeq(input: {
-    runtimeKey: string;
-    afterSeq: number;
-    limit: number;
-  }): Promise<LinearMessageLedgerRecord[]>;
-  countAfterSeq(input: { runtimeKey: string; afterSeq: number }): Promise<number>;
-  findLatest(input: { runtimeKey: string }): Promise<LinearMessageLedgerRecord | null>;
 }
