--- apps/agent/static/context/spire-unavailable.hbs
@@ -1,3 +1,0 @@
-<spire_screen>
-尖塔现在打不开（游戏服务未就绪或正在重启）。过一会儿再用 switch(id="spire") 进来，或先去别的 App。
-</spire_screen>
