+++ apps/agent/test/agent/linear-message-ledger-agent-context.test.ts
@@ -11,9 +11,6 @@ describe("LinearMessageLedgerAgentContext", () => {
       }),
       linearMessageLedgerDao: {
         insertMany,
-        listAfterSeq: vi.fn(),
-        countAfterSeq: vi.fn(),
-        findLatest: vi.fn(),
       },
       runtimeKey: "root-agent",
     });
@@ -73,9 +70,6 @@ describe("LinearMessageLedgerAgentContext", () => {
       }),
       linearMessageLedgerDao: {
         insertMany,
-        listAfterSeq: vi.fn(),
-        countAfterSeq: vi.fn(),
-        findLatest: vi.fn(),
       },
       runtimeKey: "root-agent",
     });
