+++ packages/kernel/src/contracts/llm.ts
@@ -1,4 +1,4 @@
-// provider 标识的单源在 @kagami/llm，需要 `LlmProviderId` 的代码请直接从那里导入。
+// LLM 各调用点的用途标识（落库归因用）。provider 标识（`LlmProviderId`）的单源另在 @kagami/llm。
 export type LlmUsageId =
   | "agent"
   | "contextSummarizer"
